Renamed package `too-simple-server`.
Add pypi deploy to build script
