+++ too_simple_server/__main__.py
@@ -1,7 +1,7 @@
 """Run server as module"""
 from argparse import ArgumentParser
 
-from web_server.run import main
+from too_simple_server.run import main
 
 AGP = ArgumentParser(description="Mock server with simple DB interactions")
 AGP.add_argument("--debug", action="store_true", default=False)
